Fix GPS state handling and randomize frame selection
Random frame selection is now enabled in NumpyRawFrameFileGenerator by using np.random.choice. In gps.py, sats_used is now parsed as float for consistency. In processors.py, added null checks for _gps in get_state and get_latlon to prevent errors when GPS is uninitialized.

+++ gps.py
@@ -284,7 +284,7 @@ class UsbGps(BaseGps):
 
         # ----------------------------- GNS -----------------------------
         elif stype == "GNS":
-            self._state.sats_used = int(data[6] or 0)
+            self._state.sats_used = float(data[6] or 0)
             self._state.hdop = float(data[7] or math.nan)
             self._state.alt_msl = float(data[8] or math.nan)
             self._state.geoid_sep = float(data[9] or math.nan)
